requirements.txt
python-telegram-bot
python-dotenv
requests
gTTS